Fix -Werror

+++ src/map/validation/has_proper_char.c
@@ -6,7 +6,7 @@
 /*   By: gyoon <gyoon@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/15 16:22:37 by gyoon             #+#    #+#             */
-/*   Updated: 2023/03/15 16:55:39 by gyoon            ###   ########.fr       */
+/*   Updated: 2023/03/19 21:37:47 by gyoon            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ t_bool	has_proper_char(t_map map)
 	while (y < map.size.y)
 	{
 		x = 0;
-		while (x < ft_strlen(map.map[y]) - 1)
+		while (x < (int)ft_strlen(map.map[y]) - 1)
 		{
 			if (map.map[y][x] != '0' && \
 				map.map[y][x] != '1' && \
