+++ src/map/validation/is_rectangle.c
@@ -6,7 +6,7 @@
 /*   By: gyoon <gyoon@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/15 15:55:18 by gyoon             #+#    #+#             */
-/*   Updated: 2023/03/19 20:24:36 by gyoon            ###   ########.fr       */
+/*   Updated: 2023/03/19 21:38:11 by gyoon            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ t_bool	is_rectangle(t_map map)
 	y = 0;
 	while (y < map.size.y)
 	{
-		if (map.size.x != ft_strlen(map.map[y]) - 1)
+		if (map.size.x != (int)ft_strlen(map.map[y]) - 1)
 			return (ft_false);
 		y++;
 	}
